LEGATO-ACTIVITY - Organized all activity file paths because it was driving me crazy
