+++ app/src/main/java/com/example/legatoapp/Activity/BaseActivity.java
@@ -1,4 +1,4 @@
-package com.example.legatoapp;
+package com.example.legatoapp.Activity;
 
 import androidx.appcompat.app.AppCompatActivity;
