+++ app/src/main/java/com/example/legatoapp/Activity/SettingsActivity.java
@@ -1,4 +1,4 @@
-package com.example.legatoapp;
+package com.example.legatoapp.Activity;
 
 import android.content.Intent;
 import android.content.SharedPreferences;
@@ -8,6 +8,8 @@ import android.widget.Button;
 import android.widget.ImageButton;
 
 import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.legatoapp.R;
 
 public class SettingsActivity extends AppCompatActivity {
     @Override
