+++ app/src/main/java/com/example/legatoapp/Fragment/HomeFragment.java
@@ -1,4 +1,4 @@
-package com.example.legatoapp;
+package com.example.legatoapp.Fragment;
 
 import android.os.Bundle;
 
@@ -7,6 +7,8 @@ import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import com.example.legatoapp.R;
 
 /**
  * A simple {@link Fragment} subclass.
