+++ app/src/main/java/com/example/legatoapp/models/Song.java
@@ -1,4 +1,4 @@
-package com.example.legatoapp;
+package com.example.legatoapp.models;
 
 public class Song {
     private String title;
