+++ app/src/main/res/layout/activity_edit_profile.xml
@@ -14,7 +14,7 @@
         android:layout_height="wrap_content"
         android:fitsSystemWindows="true"
         android:paddingBottom="45dp"
-        tools:context=".EditProfileActivity">
+        tools:context=".Activity.EditProfileActivity">
 
         <!-- Guideline for Centering -->
         <androidx.constraintlayout.widget.Guideline
