+++ app/src/main/res/layout/activity_user_home.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:background="@drawable/background"
     android:fitsSystemWindows="true"
-    tools:context=".userHome">
+    tools:context=".Activity.userHome">
 
 
     <FrameLayout
