+++ app/src/main/res/layout/fragment_create.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background"
-    tools:context=".CreateFragment">
+    tools:context=".Fragment.CreateFragment">
 
     <!-- Search Songs Button at the Top -->
     <Button
