+++ app/src/main/res/layout/fragment_inbox.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:padding="16dp"
-    tools:context=".InboxFragment">
+    tools:context=".Fragment.InboxFragment">
 
     <!-- Title -->
     <TextView
