+++ app/src/main/res/layout/fragment_profile.xml
@@ -14,7 +14,7 @@
         android:layout_height="wrap_content"
         android:fitsSystemWindows="true"
         android:padding="16dp"
-        tools:context=".ProfileFragment">
+        tools:context=".Fragment.ProfileFragment">
 
         <!-- Settings Button -->
         <ImageButton
